Refactor blobfuse support for AKS
